pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Bohdan <bohdan@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
